requirements.txt
discord
discoin
rethinkdb
pyyaml
